fix!(database): Cambiar datos de llaves y backup
Cambiar tipo de dato de las llaves primarias y foraneas de todas las
tablas.

Backup de la Base de datos y PGAdmin.

+++ .devcontainer/scripts/sql/Tablas.sql
@@ -2,7 +2,7 @@ DROP TABLE IF EXISTS carta_tcg, juego_de_mesa, producto, valoracion, rol, tienda
 
 -- Tabla Carta_TCG
 CREATE TABLE IF NOT EXISTS carta_tcg (
-    id_carta BIGINT PRIMARY KEY,
+    id_carta SERIAL PRIMARY KEY,
 	nombre_carta VARCHAR(32),
     tipo_carta VARCHAR(32),
 	descripcion_carta VARCHAR(500),
@@ -14,7 +14,7 @@ CREATE TABLE IF NOT EXISTS carta_tcg (
 
 -- Tabla Juego_De_Mesa
 CREATE TABLE IF NOT EXISTS juego_de_mesa (
-    id_juego_de_mesa BIGINT PRIMARY KEY, 
+    id_juego_de_mesa SERIAL PRIMARY KEY, 
     tipo_juego_de_mesa VARCHAR(32),
     stock_juegos_mesa INTEGER,
     nombre_juegos_mesa VARCHAR(64),
@@ -23,9 +23,9 @@ CREATE TABLE IF NOT EXISTS juego_de_mesa (
 
 -- Tabla Producto
 CREATE TABLE IF NOT EXISTS producto (
-    id_producto BIGINT PRIMARY KEY,
-	  id_carta BIGINT  NULL,
-	  id_juego_de_mesa BIGINT NULL,
+    id_producto SERIAL PRIMARY KEY,
+	  id_carta INT  NULL,
+	  id_juego_de_mesa INT NULL,
     tipo_producto VARCHAR(32),
     fecha_subida DATE,
     cantidad_vendidas BIGINT,
@@ -37,8 +37,8 @@ CREATE TABLE IF NOT EXISTS producto (
 );
   
 CREATE TABLE IF NOT EXISTS valoracion (
-  id_valoracion BIGINT PRIMARY KEY,
-  id_producto BIGINT,
+  id_valoracion SERIAL PRIMARY KEY,
+  id_producto INT,
   comentario_valoracion	 VARCHAR(300),
   fecha_valoracion DATE,
   puntuacion NUMERIC(2,1),
@@ -47,14 +47,14 @@ CREATE TABLE IF NOT EXISTS valoracion (
 );
 
 CREATE TABLE IF NOT EXISTS rol (
-  id_rol BIGINT PRIMARY KEY,
+  id_rol SERIAL PRIMARY KEY,
   tipo_rol CHAR(7),
   nivel_permiso SMALLINT
 );
 
 CREATE TABLE IF NOT EXISTS tienda(
-	id_tienda BIGINT PRIMARY KEY,
-	id_rol BIGINT,
+	id_tienda SERIAL PRIMARY KEY,
+	id_rol INT,
 	nombre_tienda VARCHAR(32),
 	direccion_tienda VARCHAR(255),
 	url_producto_imagen	CHAR(64),
@@ -63,9 +63,9 @@ CREATE TABLE IF NOT EXISTS tienda(
 );
 
 CREATE TABLE IF NOT EXISTS usuario (
-  id_usuario            BIGINT       PRIMARY KEY,
-  id_rol                BIGINT       ,
-  id_valoracion         BIGINT       ,
+  id_usuario            SERIAL       PRIMARY KEY,
+  id_rol                INT       ,
+  id_valoracion         INT       ,
   nombre_usuario        VARCHAR(32)  ,
   apellido              VARCHAR(32),
   email                 VARCHAR(32)  ,
@@ -78,13 +78,13 @@ CREATE TABLE IF NOT EXISTS usuario (
 );
 
 CREATE TABLE IF NOT EXISTS carrito(
-  id_carrito           BIGINT       PRIMARY KEY,
+  id_carrito           SERIAL       PRIMARY KEY,
   productos_carro      INTEGER
 );
 
 CREATE TABLE IF NOT EXISTS carrito_usuario (
-  id_usuario   BIGINT NOT NULL,
-  id_carrito   BIGINT NOT NULL,
+  id_usuario   INT NOT NULL,
+  id_carrito   INT NOT NULL,
  
   PRIMARY KEY (id_usuario, id_carrito),
   FOREIGN KEY (id_usuario) REFERENCES usuario(id_usuario),
@@ -92,8 +92,8 @@ CREATE TABLE IF NOT EXISTS carrito_usuario (
 );
 
 CREATE TABLE IF NOT EXISTS medio_de_pago (
-  id_medio             BIGINT       PRIMARY KEY,
-  id_carrito           BIGINT,
+  id_medio             SERIAL PRIMARY KEY,
+  id_carrito           INT,
   nombre_medio         VARCHAR(32),
   url_medio            CHAR(64),
 
@@ -101,17 +101,17 @@ CREATE TABLE IF NOT EXISTS medio_de_pago (
 );
 
 CREATE TABLE IF NOT EXISTS boleta (
-  id_boleta               BIGINT       PRIMARY KEY,
+  id_boleta               SERIAL PRIMARY KEY,
   direccion_boleta        VARCHAR(255),
   valor_boleta            INTEGER,
   numero_boleta           BIGINT
 );
 
 CREATE TABLE IF NOT EXISTS compra (
-  id_compra              BIGINT       PRIMARY KEY,
-  id_carrito             BIGINT,
-  id_boleta              BIGINT,
-  id_medio               BIGINT,
+  id_compra              SERIAL PRIMARY KEY,
+  id_carrito             INT,
+  id_boleta              INT,
+  id_medio               INT,
   cantidad_comprada      INTEGER,
   fecha_compra           DATE,
   fecha_emision_boleta   DATE,
@@ -122,8 +122,8 @@ CREATE TABLE IF NOT EXISTS compra (
 );
 
 CREATE TABLE IF NOT EXISTS tienda_producto (
-  id_tienda BIGINT,
-  id_producto BIGINT,
+  id_tienda SERIAL,
+  id_producto INT,
 
   PRIMARY KEY (id_tienda, id_producto),
   FOREIGN KEY (id_tienda) REFERENCES tienda(id_tienda),
@@ -131,8 +131,8 @@ CREATE TABLE IF NOT EXISTS tienda_producto (
 );
 
 CREATE TABLE IF NOT EXISTS boleta_producto (
-  id_boleta BIGINT,
-  id_producto BIGINT,
+  id_boleta SERIAL,
+  id_producto INT,
 
   PRIMARY KEY (id_boleta, id_producto),
   FOREIGN KEY (id_boleta) REFERENCES boleta(id_boleta),
@@ -140,14 +140,14 @@ CREATE TABLE IF NOT EXISTS boleta_producto (
 );
 
 CREATE TABLE IF NOT EXISTS listadeseos (
-  id_lista_deseos BIGINT PRIMARY KEY,
+  id_lista_deseos SERIAL PRIMARY KEY,
   productos_en_descuento INTEGER,
   cantidad_productos INTEGER
 );
 
 CREATE TABLE IF NOT EXISTS listadeseos_usuario (
-  id_usuario BIGINT ,
-  id_lista_deseos BIGINT,
+  id_usuario SERIAL ,
+  id_lista_deseos INT,
 
   PRIMARY KEY (id_usuario, id_lista_deseos),
   FOREIGN KEY (id_usuario) REFERENCES usuario(id_usuario),
@@ -155,8 +155,8 @@ CREATE TABLE IF NOT EXISTS listadeseos_usuario (
 );
 
 CREATE TABLE IF NOT EXISTS carrito_producto (
-  id_carrito BIGINT ,
-  id_producto BIGINT ,
+  id_carrito SERIAL ,
+  id_producto INT ,
 
   PRIMARY KEY (id_carrito, id_producto),
   FOREIGN KEY (id_carrito) REFERENCES carrito(id_carrito),
@@ -164,8 +164,8 @@ CREATE TABLE IF NOT EXISTS carrito_producto (
 );
 
 CREATE TABLE IF NOT EXISTS listadeseos_producto(
-	id_lista_deseos BIGINT,
-	id_producto BIGINT    ,
+	id_lista_deseos INT,
+	id_producto INT    ,
 
 	 PRIMARY KEY (id_lista_deseos, id_producto),
 	 FOREIGN KEY (id_lista_deseos) REFERENCES listadeseos(id_lista_deseos),
